refactor(layerChooser): migrate utility function getBackgroundsAndMasks
